+++ src/model/project.rs
@@ -3,10 +3,10 @@
 //! `model::project` is a model for the Ruddr Project object. This module is not publically accessible, but the structs and members are public for reading from `interface::project` returns.
 //! [API Documentation](https://ruddr.readme.io/reference/project-object)
 use crate::model::{enums, shared, types};
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 /// Model for Projects used with List operations.
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Projects {
     pub results: Vec<Project>,
@@ -14,7 +14,7 @@ pub struct Projects {
 }
 
 /// Model for Project used with Read operations.
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Project {
     pub id: types::UUID,
@@ -65,7 +65,7 @@ pub struct Project {
     pub monthly_budget: Option<MonthlyBudget>,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Budget {
     pub revenue: i64,
@@ -77,7 +77,7 @@ pub struct Budget {
     pub non_billable_hours: i64,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct MonthlyBudget {
     pub revenue: i64,
@@ -90,7 +90,7 @@ pub struct MonthlyBudget {
 }
 
 // custom types: enum
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "snake_case")]
 pub enum BillingType {
     #[serde(rename = "tm")]
@@ -100,14 +100,14 @@ pub enum BillingType {
     NonBillable,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "snake_case")]
 pub enum RevenueRecognitionMethod {
     Invoiced,
     Manual,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "snake_case")]
 pub enum BudgetMode {
     Summary,
@@ -115,14 +115,14 @@ pub enum BudgetMode {
     Aggregated,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "snake_case")]
 pub enum MonthlyBudgetMode {
     Summary,
     Detailed,
 }
 
-#[derive(PartialEq, Deserialize, Debug)]
+#[derive(PartialEq, Deserialize, Serialize, Debug)]
 #[serde(rename_all = "snake_case")]
 pub enum RecordStatus {
     Active,