+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.1.1 (Next)
 - Restore model deserialization validation.
 - Derive `Clone` for custom types.
+- Derive `Serialize` for struct models.
 
 ### 1.1.0
 - Add support for project roles.